Moves timer and ReconnectTask in the wrapper.
This way we will have more rapid reconnects as there will be a thread
for every reconnect task, and not a single one for all tasks.
